Fix #17: Hide sub-categories column when sub-categories are not shown
Modified the createFinancialOverview function to dynamically hide or show the sub-categories column based on the user preference. When sub-categories are not to be shown (showSubCategories is false), column 3 (Sub-Category) is now hidden using sheet.hideColumn(3). When sub-categories are to be shown (showSubCategories is true), the column is made visible using sheet.showColumn(3).